Read description for update details
index.html - updated the positioning of my CSS and JavaScript tags in order for correct loading of the DOM and objects.

stylesheets - included a few divs into the message container for other functions handling joinchannel() and other functions for chat box

JavaScript files - implemented features for closing chat box when to-top is clicked, implimented new functions for handling logout of agoraRTM and leaving a channel. also included more conditionals for icons in resume section so when a user taps or clicks the icons will not translate.