Fixing loading problems with lowend devices
Transform scales on load and registering variable information on low end devices for the Icon slider in resume.js was a problem. The loading of CSS was not complete before JavaScript registered even on low end devices and chrome dev tools throttling set. All script tags are set to defer at the bottom of HTML. This is an on going issue to create a more dynamic width variable for Icons in resume.js. Problem is still not solved completely.