Updating functions and styles
creating a members count with channel.getMembers() RTM built in function. And forcing a page reload after leaving the channel. Can't find a way for Agora RTM to stop creating a new instance even after logout and channel.leave(). One last implementation was adding in a logical operator || to watch for a capitol "L" when typing in "Leave chat". One of the most common inputs, given mobile auto corrects

in resume.js I implemented styles for better experience in my personal taste and in sass, I updated some styling on send button as well as canceling a grid layout on @media max-height.