Creating another IIFE function for mobile devices read more in description
I considered only changing the layout on mobile to allow a smaller javascript file, but ultimately I am here to show my skills and knowleddge and the goal is to do that with JavaScript so I chose to make the page more dynamic with JS instead of SASS.

the mobile IIFE is checking for a double tap. The reason is because I chose to not update the dom with sass, instead update it dynamically in two different ways based of screen width. double touch for mobile runs one thing while a button click on computers runs another. I cannot mix them up and needed a function to handle double touches instead of double clicks so PC users did not get unwanted mobile effects