Updating more functionality and experience
Agora: Integrated an element and functions to update message notifications. along with adding user name when joined to welcome message, a condition to only append one leave box, and creating a default channel

resume.js: Successfully adding the conditional for no translation of the skill icons if the event is not a swipe greater than 75px

index/styles: adding in elements and their corresponding styles for user experience.