Finally accomplishing a typing indication from other user. See details
Because AgoraRTM does not provide a function that listens to user typing I needed to create my own. The only way I could allow this to work is by adding two event listeners to the form and updating the message3 variable multiple times. All is well, except for one exception. To send a signal through to the other user, it was necessary for me to call the channel.on('ChannelMessage') function provided by agora. This means that now each letter that is typed a new message is sent through the API and delivered. The function basically turns that message into the typing animation for the other end, and rejects adding it to the dom for the user typing. Else... If the keydown event === 'Enter' a condition is watching before sending the message and returns early before the send message function can fully execute allowing for the forms submit event to take back over. The issue comes with performance setbacks and massive unneccessary data transfers. The code will be commented out when the site begins to experiences traffic.