Adding some styling and creating dynamic measurements for hidden menu
created better styling for projects section

I also added in dynamic measurements for hidden menu regarding the conditions met for drag and drop.

the last implementation is for the card banners in my projects section. I created a function to automatically update the banner number when new projects are added.